refactor(apps): split client setup — configureClient into plugin runtime entry
Per review: give the plugin two entry points and share only the universal
half of the old hand-written client index.

- @kestra-io/hey-api-plugin now ships two entries: "." (the codegen plugin,
  generation-time) and "./runtime" (createConfigureClient — the universal
  axios setup: Content-Type/Accept fixes, QueryFilter serializer, error
  normalization). axios is an optional peer dep of the runtime entry.
- The OSS SDK sources block-1 from ./runtime (bundled into dist via
  noExternal, so no new runtime dependency) and keeps the app-only
  singleton (useClient/setMockClient) in client-setup.ts. createConfigureClient
  no longer holds module state; the app records the built instance.
- build:sdk builds the plugin first (fast, Java-free).

Note: the plugin's package.json gained axios as a dev/peer dep; the
committed lockfile needs regenerating in a healthy-npm environment (this
sandbox's npm arborist crashes on install/lockfile regen).
