+++ ui/package.json
@@ -14,7 +14,7 @@
         "build": "vite build --emptyOutDir",
         "build:design": "npm run build --workspace=@kestra-io/design-system",
         "build:topology": "npm run build --workspace=@kestra-io/topology",
-        "build:sdk": "npm run build --workspace=@kestra-io/kestra-sdk",
+        "build:sdk": "npm run build --workspace=@kestra-io/hey-api-plugin && npm run build --workspace=@kestra-io/kestra-sdk",
         "generate:openapi-spec": "cd .. && ./gradlew :webserver:generateOpenapiSpec",
         "generate:sdk": "npm run generate:openapi-spec && npm run build --workspace=@kestra-io/hey-api-plugin && npm run generate:openapi --workspace=@kestra-io/kestra-sdk && npm run build:sdk",
         "preview": "vite preview",
